Static or Dynamic Raft node list
This allows to choose between a statically, precompiled list of
retry_join nodes to configure the Raft backend (vault_backend: raft) or
a dynamically generated list that is generated using Ansible facts
(vault_backend: raft_dynamic).

The reason for this backward compatible change is that having a dynamic
option only (vault_backend: raft_dynamic) does not work for playbooks
that run against a single node only, for instance, to perform an upgrade
(batch size one).