Define vault_api_addr and vault_cluster_addr using vault_hostname (when available)
Before this commit, vault_api_addr and vault_cluster_addr were defined
using the IPv4 address of the default interface of the host. Moreover, a
variable named `vault_hostname` is used to run the reachability check
but is undocumented.

This change documents `vault_hostname` and leverages it to set the value
of `vault_api_addr` and `vault_cluster_addr` when it's available.
Otherwise, these vars are defined as before.