Fix leader_api_addr in retry_join stanza
Prior to this fix, leader_api_addr took the value of vault_api_addr for
other nodes defined in the cluster. However, if this value is templated,
the template string is used instead of the resolved value.

To fix that, this commit sets a fact containing the resolved value right
before writing the config file to disk.